requirements.txt
click==8.1.8
markdown-it-py==3.0.0
mdurl==0.1.2
pygments==2.19.1
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
typer==0.16.0
typing-extensions==4.14.0
